chore: rename screen-only.ts composable
